misc/HtmlTestSuite/results/test-1112.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>HTML Test: focus - default value of tabindex</title>
<link href="http://www.intel.com/" rel="author" title="Intel"/>
<link href="https://html.spec.whatwg.org/multipage/#sequential-focus-navigation-and-the-tabindex-attribute" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><div id="log"/>
<button id="test1">TEST1</button>
<div id="test2">TEST2</div>
<script>

test(function() {
  assert_equals(document.getElementById("test1").tabIndex, 0, "The value of tabIndex attribute should be 0.");
}, "The default value of tabIndex attribute must be 0 for elements that are focusable");

test(function() {
  assert_equals(document.getElementById("test2").tabIndex, -1, "The value of tabIndex attribute should be -1.");
}, "The default value of tabIndex attribute must be -1 for elements that are not focusable");

</script>
</body></html>
